v5.3.6: Giveaways can now be ended manually; Giveaways, Quickpolls, Polls and Signups have now a hint on what the X-reaction is for; MediaOnlyChannels now properly find URLs; Bug Fixes.

+++ src/main/java/utilities/MessageHandler.java
@@ -15,7 +15,7 @@ public class MessageHandler {
         channel.sendMessage(message).queue(sentMessage -> new Timer().schedule(new TimerTask() {
             @Override
             public void run() {
-                sentMessage.delete().queue();
+                if (sentMessage != null) sentMessage.delete().queue();
             }
         }, cooldown));
     }
